requirement.txt
alembic==1.13.3
async-timeout==4.0.3
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
docker==7.1.0
eventlet==0.37.0
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
redis==4.5.5
requests==2.32.3
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
